requirements.txt
streamlit
opencv-python-headless
numpy
torch
torchvision
timm
Pillow
python-magic
reportlab
ollama
librosa
transformers